+++ src/Main/WPContextRequester.php
@@ -1,0 +1,91 @@
+<?php
+
+declare(strict_types=1);
+
+namespace luisdeb\Woncer\Main;
+
+/**
+ * Helper class to provide with HTTP context for a Nonce request.
+ *
+ * @since v0.0.2
+ */
+class WPContextRequester implements \ArrayAccess
+{
+    /**
+     * The HTTP request data based on method.
+     *
+     * @var array $httpRequest
+     */
+    private $httpRequest = [];
+
+    /**
+     * The HTTP request method.
+     *
+     * @var string $httpMethod
+     */
+    private $httpMethod = '';
+
+    /**
+     * Class constructor covering basic parameters.
+     * Initializes the action, name and token properties.
+     *
+     */
+    public function __construct()
+    {
+        $inputMethod = filter_input(INPUT_SERVER, 'REQUEST_METHOD', FILTER_SANITIZE_SPECIAL_CHARS);
+        $this->httpMethod = !isset($inputMethod) ? null : $inputMethod;
+        
+        if(isset($this->httpMethod) && 
+        strtoupper($this->httpMethod) === 'GET' ||
+        strtoupper($this->httpMethod) === 'POST')
+        {
+            $this->httpRequest = array_merge($_GET, $_POST);
+        }
+    }
+
+    /**
+     * {@inheritDoc} From ArrayAccess implementation 
+     */
+    public function offsetExists($offset): bool
+    {
+        return isset($this->httpRequest[$offset]);
+    }
+
+    /**
+     * {@inheritDoc} From ArrayAccess implementation 
+     * 
+     * Offset getter method
+     */
+    public function offsetGet($offset)
+    {
+        return isset($this->httpRequest[$offset]) ? $this->httpRequest[$offset] : null;
+    }
+
+    /**
+     * {@inheritDoc} From ArrayAccess implementation 
+     * 
+     * Offset setter method is disabled for security reasons
+     */
+    public function offsetSet($offset, $value)
+    {
+        throw new Exception("Request parameters are immutable"); 
+    }
+
+    /**
+     * {@inheritDoc} From ArrayAccess implementation 
+     */
+    public function offsetUnset($offset)
+    {
+        throw new Exception("Request parameters are immutable"); 
+    }
+
+    /**
+     * getter method for the http request associative array
+     * 
+     * @return the request array
+     */
+    public function httpRequest(): array
+    {
+        return $this->httpRequest;
+    }
+}