+++ src/Main/WPNonceChecker.php
@@ -4,15 +4,15 @@ declare(strict_types=1);
 
 namespace luisdeb\Woncer\Main;
 
-use luisdeb\Woncer\Main\WPNonce as WPNonce;
+use luisdeb\Woncer\Main\WPNonceAbstract as WPNonceAbstract;
 
 /**
- * This class handles the logic for Nonce verification.
- * It implements different methods depending on Nonce context.
+ * General-purpose Nonce verification class.
+ * It implements different verifier methods depending on Nonce context.
  *
  * @see https://codex.wordpress.org/Wordpress_Nonce_Implementation
  */
-class WPNonceChecker extends WPNonce
+class WPNonceChecker extends WPNonceAbstract
 {
     /**
      * The current WordPress `check admin referer function name`.
@@ -36,6 +36,57 @@ class WPNonceChecker extends WPNonce
     const VERIFY_NONCE_FUNCTION_NAME = 'wp_verify_nonce';
 
     /**
+     * Class constructor covering basic parameters.
+     * Initializes the action, name and token properties.
+     *
+     * @param string $action
+     * @param string $name
+     */
+    public function __construct(string $action, string $name)
+    {
+        parent::__construct($action, $name);
+    }
+
+    /**
+     * Verifies the nonce passed in some other context
+     *
+     * @see https://codex.wordpress.org/Function_Reference/wp_verify_nonce
+     *
+     * @param WPContextRequester $httpContext | the HTTP request.
+     *
+     * @var WPContextRequester $httpContext | object instance containing the request data.
+     * @var string $nonce | stores the nonce name from request.
+     * @var string $nonceAction | stores the nonce action from request.
+     *
+     * @return int $result | value `0` if nonce is invalid.
+     *                     | value `1` if nonce was generated in the past 12 hours or less.
+     *                     | value `2` if nonce was generated between 12 and 24 hours ago.
+     *
+     * wp_verify_nonce returns a boolean `false`. In this function we return an int `0` as `false`,
+     * as multiple return types are not supported yet, and will not be until PHP 8.0 is released.
+     *
+     * @see https://wiki.php.net/rfc/union_types_v2
+     *
+     */
+    public function verifyNonce(WPContextRequester $httpContext = null): int
+    {
+        $result = 0;
+        $httpContext = (!$httpContext) ? new WPContextRequester() : $httpContext;
+        $nonce = $httpContext->offsetExists($this->name()) ? $httpContext[$this->name()] : '';
+        $nonceAction = $httpContext->offsetExists($this->action()) ? $httpContext[$this->action()] : '';
+
+        if (!is_string($nonce) || !is_string($nonceAction)) {
+            return 0;
+        }
+
+        if (function_exists(self::VERIFY_NONCE_FUNCTION_NAME)) {
+            $result = wp_verify_nonce($nonce, $nonceAction);
+        }
+
+        return (!$result) ? 0 : $result;
+    }
+
+    /**
      * Tests if the nonce is valid or if the request was referred from an admin.
      *
      * @see https://codex.wordpress.org/Function_Reference/check_admin_referer
@@ -49,7 +100,7 @@ class WPNonceChecker extends WPNonce
      */
     public function checkAdminReferer(
         string $action = null,
-        string $queryArg = null
+        string $queryArg = '_wpnonce'
     ): bool {
 
         $result = false;
@@ -89,40 +140,5 @@ class WPNonceChecker extends WPNonce
         }
 
         return $result;
-    }
-
-    /**
-     * Verifies the nonce passed in some other context
-     *
-     * @see https://codex.wordpress.org/Function_Reference/wp_verify_nonce
-     *
-     * @param string $nonce     | the nonce to verify.
-     * @param string $action    | the context for the nonce.
-     *
-     * @var string $nonceAction | takes the local action value in case $action is not set.
-     *
-     * @return int $result | value `0` if nonce is invalid.
-     *                     | value `1` if nonce was generated in the past 12 hours or less.
-     *                     | value `2` if nonce was generated between 12 and 24 hours ago.
-     *
-     * wp_verify_nonce returns a boolean `false`. In this function we return an int `0` as `false`,
-     * as multiple return types are not supported yet, and will not be until PHP 8.0 is released.
-     *
-     * @see https://wiki.php.net/rfc/union_types_v2
-     *
-     */
-    public function verifyNonce(
-        string $nonce,
-        string $action = null
-    ): int {
-
-        $result = 0;
-        $nonceAction = (!$action) ? $this->action() : $action;
-
-        if (function_exists(self::VERIFY_NONCE_FUNCTION_NAME)) {
-            $result = wp_verify_nonce($nonce, $nonceAction);
-        }
-
-        return (!$result) ? 0 : $result;
     }
 }